master fn

+++ bin/mk.sh
@@ -3,23 +3,36 @@
 # Развернуть указанный сниппет в текущий каталог
 
 export pkg=$2
-export pypkg=`echo "$pkg" | sed 's/-/_/g'`
+export pypkg=`echo "$pkg" | sed 's/-/_/g' | sed 's/python_//g'`
 snippet=$RIG_RC/snippet/$1
 
+echo "***************************************"
+echo "** $snippet"
+echo "***************************************"
+
 replace_vars() {
-	perl -npe 's!\{\{(\w+)\}\}!$ENV{$1} // "???"!ge'
+	perl -npe 's!\{\{(\w+)\}\}!$ENV{$1} // $&!ge'
 }
 
 cp_path() {
 	path=$1
-	to=`echo $1 | sed 's/^$snippet//'`
+	to=`echo $pkg${1#$snippet}`
 	to=`echo $to | replace_vars`
+
+	echo $to
+	
+	if [ -e $to ]; then
+		echo "Файл существует."
+		exit 1
+	fi
 
 	if [ -f $path ]; then
 		cat $path | replace_vars > $to
-	else
+	elif [ -d $path ]; then
 		mkdir -p $to
+	else
+		echo "Странный путь `$path`"
 	fi 
 }
 
-find $snippet -exec cp_path "{}" \;
+find $snippet | while read file; do cp_path "$file"; done
